requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
beautifulsoup4==4.6.0
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
decorator==4.2.1          # via ipython, traitlets
django-extensions==1.9.9
django==2.0.1
factory-boy==2.9.2
faker==0.8.10
idna==2.6                 # via requests
ipdb==0.10.3
ipython-genutils==0.2.0   # via traitlets
ipython==6.2.1
jedi==0.11.1              # via ipython
parso==0.1.1              # via jedi
pexpect==4.3.1            # via ipython
pickleshare==0.7.4        # via ipython
prompt-toolkit==1.0.15    # via ipython
psycopg2==2.7.3.2
ptyprocess==0.5.2         # via pexpect
pygments==2.2.0           # via ipython
python-dateutil==2.6.1    # via faker
pytz==2017.3              # via django
requests==2.18.4
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via django-extensions, faker, prompt-toolkit, python-dateutil, traitlets
text-unidecode==1.1       # via faker
traitlets==4.3.2          # via ipython
typing==3.6.2             # via django-extensions
urllib3==1.22             # via requests
wcwidth==0.1.7            # via prompt-toolkit
